+++ CHANGELOG.md
@@ -1,9 +1,20 @@
-# Change Log
+# Changelog
 
 All notable changes to the "laragon-terminal-config" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+This project adheres to [Semantic Versioning](http://semver.org/).
+
+The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 
-- Initial release
+### Added
+- Customizable setting `laragonTerminalConfig.laragonPath` to specify the path to the Laragon Cmder executable.
+- Boolean setting `laragonTerminalConfig.setAsDefault` to control whether Laragon should be set as the default terminal.
+- Command `laragon-terminal-config.setProfile` to update the terminal configuration manually from the Command Palette.
+- Automatic updating of the Windows terminal profile upon extension activation if the setting is enabled.
+
+## [0.0.1] - YYYY-MM-DD
+
+### Initial Release
+- Initial release of the Laragon Terminal Config extension.